Merge bazelbuild/bazel@master

+++ src/main/java/com/google/devtools/build/lib/sandbox/AsynchronousTreeDeleter.java
@@ -20,6 +20,8 @@ import static com.google.common.base.Preconditions.checkState;
 import com.google.common.flogger.GoogleLogger;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
 import com.google.devtools.build.lib.exec.TreeDeleter;
+import com.google.devtools.build.lib.profiler.Profiler;
+import com.google.devtools.build.lib.profiler.SilentCloseable;
 import com.google.devtools.build.lib.vfs.Path;
 import java.io.IOException;
 import java.util.concurrent.LinkedBlockingQueue;
@@ -37,7 +39,10 @@ import javax.annotation.Nullable;
  * precious resources that could otherwise be used for the build itself. But when the build is
  * finished, this number should be raised to quickly go through any pending deletions.
  */
-class AsynchronousTreeDeleter implements TreeDeleter {
+public class AsynchronousTreeDeleter implements TreeDeleter {
+
+  public static final String MOVED_TRASH_DIR = "_moved_trash_dir";
+
   private static final GoogleLogger logger = GoogleLogger.forEnclosingClass();
 
   private final AtomicInteger trashCount = new AtomicInteger(0);
@@ -47,8 +52,10 @@ class AsynchronousTreeDeleter implements TreeDeleter {
 
   private final Path trashBase;
 
+  private boolean trashBaseCreated = false;
+
   /** Constructs a new asynchronous tree deleter backed by just one thread. */
-  AsynchronousTreeDeleter(Path trashBase) {
+  public AsynchronousTreeDeleter(Path trashBase) {
     logger.atInfo().log("Starting async tree deletion pool with 1 thread");
 
     ThreadFactory threadFactory =
@@ -82,12 +89,19 @@ class AsynchronousTreeDeleter implements TreeDeleter {
 
   @Override
   public void deleteTree(Path path) throws IOException {
+    if (!trashBaseCreated) {
+      trashBase.createDirectory();
+      trashBaseCreated = true;
+    }
+    if (!path.exists()) {
+      return;
+    }
     Path trashPath = trashBase.getRelative(Integer.toString(trashCount.getAndIncrement()));
     path.renameTo(trashPath);
     checkNotNull(service, "Cannot call deleteTree after shutdown")
         .execute(
             () -> {
-              try {
+              try (SilentCloseable c = Profiler.instance().profile("trashPath.deleteTree")) {
                 trashPath.deleteTree();
               } catch (IOException e) {
                 logger.atWarning().withCause(e).log(
@@ -103,5 +117,9 @@ class AsynchronousTreeDeleter implements TreeDeleter {
       service.shutdown();
       service = null;
     }
+  }
+
+  public Path getTrashBase() {
+    return trashBase;
   }
 }
